+++ java/fory-core/src/main/java/org/apache/fory/serializer/MetaSharedSerializer.java
@@ -40,8 +40,8 @@ import org.apache.fory.resolver.TypeResolver;
 import org.apache.fory.type.Descriptor;
 import org.apache.fory.type.DescriptorGrouper;
 import org.apache.fory.type.Generics;
+import org.apache.fory.util.DefaultValueUtils;
 import org.apache.fory.util.Preconditions;
-import org.apache.fory.util.ScalaDefaultValueUtils;
 import org.apache.fory.util.record.RecordInfo;
 import org.apache.fory.util.record.RecordUtils;
 
@@ -79,7 +79,7 @@ public class MetaSharedSerializer<T> extends AbstractObjectSerializer<T> {
   private final ClassInfoHolder classInfoHolder;
   private final SerializationBinding binding;
   private final boolean hasScalaDefaultValues;
-  private final ScalaDefaultValueUtils.ScalaDefaultValueField[] scalaDefaultValueFields;
+  private final DefaultValueUtils.DefaultValueField[] scalaDefaultValueFields;
 
   public MetaSharedSerializer(Fory fory, Class<T> type, ClassDef classDef) {
     super(fory, type);
@@ -115,13 +115,14 @@ public class MetaSharedSerializer<T> extends AbstractObjectSerializer<T> {
     }
     binding = SerializationBinding.createBinding(fory);
     if (fory.getConfig().isScalaOptimizationEnabled()) {
-      hasScalaDefaultValues = ScalaDefaultValueUtils.hasScalaDefaultValues(type);
+      hasScalaDefaultValues =
+          DefaultValueUtils.getScalaDefaultValueSupport().hasDefaultValues(type);
       scalaDefaultValueFields =
-          ScalaDefaultValueUtils.buildScalaDefaultValueFields(
-              fory, type, descriptorGrouper.getSortedDescriptors());
+          DefaultValueUtils.getScalaDefaultValueSupport()
+              .buildDefaultValueFields(fory, type, descriptorGrouper.getSortedDescriptors());
     } else {
       hasScalaDefaultValues = false;
-      scalaDefaultValueFields = new ScalaDefaultValueUtils.ScalaDefaultValueField[0];
+      scalaDefaultValueFields = new DefaultValueUtils.DefaultValueField[0];
     }
   }
 
@@ -215,7 +216,7 @@ public class MetaSharedSerializer<T> extends AbstractObjectSerializer<T> {
 
     // Set default values for missing fields in Scala case classes
     if (hasScalaDefaultValues) {
-      ScalaDefaultValueUtils.setScalaDefaultValues(obj, scalaDefaultValueFields);
+      DefaultValueUtils.setDefaultValues(obj, scalaDefaultValueFields);
     }
 
     return obj;
